pyproject.toml
[tool.poetry]
name = "caribou"
version = "0.1.0"
description = ""
license = "Apache-2.0"
authors = [
    "Viktor Gsteiger <vgsteiger@student.ethz.ch>",
    "Daniel Long <pinlong@student.ubc.ca>",
    "Mohammad Shahrad <mohammad.shahrad@ubc.ca>",
]
readme = "README.md"
repository = "https://github.com/ubc-cirrus-lab/caribou"

[tool.poetry.scripts]
caribou = "caribou.deployment.client.caribou:main"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
boto3 = "^1.34.43"
jsonschema = "^4.21.1"
googlemaps = "^4.10.0"
google-auth = "^2.28.0"
numpy = "^1.26.4"
pandas = "^2.2.0"
click = "^8.1.7"
pydantic = "^2.6.4"
bs4 = "^0.0.2"
pytz = "^2024.1"

[tool.poetry.extras]
# This can be used for a more fine-grained dependency management in the future

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
pytest = "^7.4.4"
coverage = "^7.4.1"
mypy = "^1.8.0"
types-pyyaml = "^6.0.12.12"
types-jsonschema = "^4.21.0.0"
isort = "^5.13.2"
setuptools = "^69.1.0"
black = "^23.12.1"
pylint = "^3.0.3"
pytest-cov = "^4.1.0"
types-requests = "^2.31.0"
seaborn = "^0.13.1"
httpie = "^3.2.2"
gtts = "^2.5.1"
profanity = "^1.1"
pydub = "^0.25.1"
scikit-learn = "^1.4.0"
pillow = "^10.2.0"
pip-install-test = "^0.5"
dna-features-viewer = "^3.1.3"
networkx = "^3.2.1"
types-pytz = "^2024.1.0.20240203"
ipykernel = "^6.29.3"
tqdm = "^4.66.2"
statsmodels = "^0.14.1"
opencv-python = "^4.9.0.80"
torchvision = "^0.17.1"
pyaes = "^1.6.1"
markdown = "^3.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 3

[tool.coverage.run]
branch = true

[tool.coverage.html]
directory = "coverage_html_report"

[tool.mypy]
python_version = "3.10"
mypy_path = "caribou"
check_untyped_defs = true
disallow_untyped_defs = true
follow_imports = "skip"
ignore_missing_imports = true
pretty = true
exclude = [
    "caribou/tests",
    "caribou/common/setup",
    "caribou/common/models/remote_client/mock_remote_client.py",
    ]

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
extend_skip_glob = [
    "**/*/__init__.py",
    "**/tests/**",
    "**/*.pyi",
    "**/*.pyc"
    ]

[tool.black]
line-length = 120
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | ./caribou/deployment/client/cli/template/app.py
)/
'''

[tool.pylint]
max-line-length = 120
disable = [
      "C0116",  # Missing function or method docstring
      "C0114",  # Missing module docstring
      "C0115",  # Missing class docstring
      "W0621",  # Redefining name %r from outer scope (line %s) (we need this for our logging)
      "R0913",  # Too many arguments
      "R0801",  # Similar lines in %s files (we will have this because of how serverless functions need to be self-contained)
      "W0511",  # TODO
      "R0903",  # Too few public methods (we need this for our dataclasses)
      "R0914",  # Too many local variables (too noisy)
    ]
ignore = [
    "tests",
    "caribou/deployment/client/cli/template"
    ]

[tool.pytest.ini_options]
testpaths = ["caribou/tests"]
addopts = [
      "--cov=caribou",
      "--cov-report=html",
    ]